refactor: review sync service initialisation
